% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{avoid_new_lines}
\alias{avoid_new_lines}
\title{Avoid new lines in sprintf output. Function helps to avoid line lenght
limits without affecting sprintf output}
\usage{
avoid_new_lines(fmt)
}
\arguments{
\item{fmt}{string, formatted as sprintf input}
}
\value{
string
}
\description{
Avoid new lines in sprintf output. Function helps to avoid line lenght
limits without affecting sprintf output
}
\examples{
sprintf(avoid_new_lines(
  "Lorem ipsum dolor sit amet, \%s adipiscing elit, sed do eiusmod
  tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim
  veniam."
), "consectetur")

}
