% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_linter.R
\name{.getLoadedPackages}
\alias{.getLoadedPackages}
\title{Collect package names from library()/require() calls in code}
\usage{
.getLoadedPackages(code)
}
\arguments{
\item{code}{string R code (may span multiple chunks).}
}
\value{
character vector of package names.
}
\description{
Collect package names from library()/require() calls in code
}
\keyword{internal}
