% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{get_github_hostname}
\alias{get_github_hostname}
\title{this function help figuring out which GitHub domain should be used
github.roche.com will be chosen if available
otherwise github.com}
\usage{
get_github_hostname()
}
\description{
this function help figuring out which GitHub domain should be used
github.roche.com will be chosen if available
otherwise github.com
}
