% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{lintPkg}
\alias{lintPkg}
\title{Lint a package.}
\usage{
lintPkg(pkg_dir = ".")
}
\arguments{
\item{pkg_dir}{String of path to package directory.
Defaults to the current directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Convenience function which uses the gDR platform linting configuration.
}
