% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{lintPkgDirs}
\alias{lintPkgDirs}
\title{Lint select subdirectories in a package directory.}
\usage{
lintPkgDirs(pkg_dir = ".", shiny = FALSE)
}
\arguments{
\item{pkg_dir}{String of path to package directory.}

\item{shiny}{Boolean of whether or not a \code{shiny} directory should
also be lint.
Defaults to the current directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Lint select subdirectories in a package directory.
}
\details{
Will look for files in the following directories:
\code{"R"}, \code{"tests"}, and conditionally \code{"inst/shiny"}
if \code{shiny} is \code{TRUE}.
}
\examples{
lintPkgDirs(pkg_dir= "gDRstyle/tst_pkgs/dummy_pkg")

}
