% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter.R
\name{lintPkgInDir}
\alias{lintPkgInDir}
\title{Lint select subdirectories in a package directory.}
\usage{
lintPkgInDir(pkg_dir = ".")
}
\arguments{
\item{pkg_dir}{String of path to package directory.
Defaults to the current directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Lint select subdirectories in a package directory.
}
