% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_custom.R
\name{roxygen_author_tag_linter}
\alias{roxygen_author_tag_linter}
\title{roxygen_author_tag_linter}
\usage{
roxygen_author_tag_linter(tag = "@author")
}
\arguments{
\item{tag}{character (default \code{@author})}
}
\value{
\code{\link[lintr]{Linter}}
}
\description{
Check that function has documented specific tag in Roxygen skeleton (default \code{@author}).
}
\examples{
\dontrun{
  linters_config <- lintr::with_defaults(
  line_length_linter = lintr::line_length_linter(120),
  roxygen_author_tag_linter = roxygen_author_tag_linter()
  )
}
}
