% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_custom.R
\name{roxygen_tag_linter}
\alias{roxygen_tag_linter}
\title{roxygen_tag_linter}
\usage{
roxygen_tag_linter(tag = "@author")
}
\arguments{
\item{tag}{character (default \code{@author})}
}
\value{
linter class function
}
\description{
Check that function has documented specific tag in Roxygen
skeleton (default \code{@author}).
}
\examples{
 linter_helper_function <- if (packageVersion("lintr") >= "3.0.0") {
   lintr::linters_with_defaults
 } else {
   lintr::with_defaults
 }
linters_config <- linter_helper_function(
  line_length_linter = lintr::line_length_linter(120),
  roxygen_tag_linter = roxygen_tag_linter()
)

}
\author{
Kamil Foltynski <kamil.foltynski@contractors.roche.com>
}
