% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{checkDependencies}
\alias{checkDependencies}
\title{Check for alignment of package dependencies across Rplatform and package specifications.}
\usage{
checkDependencies(dep_path, desc_path)
}
\arguments{
\item{dep_path}{String of path to the rplatform \code{dependencies.yaml} file.}

\item{desc_path}{String of the path to the package \code{DESCRIPTION} file.}

\item{pkg_dir}{String of path to package directory. 
Defaults to the current directory.}
}
\value{
\code{NULL} invisibly.
}
\description{
Check the package depedency version specifications in the \code{rplatform/dependencies.yaml} and \code{DESCRIPTION}.
}
\details{
This function is used for its side effects in the event that there are dependency clashes.
}
