% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPackage}
\alias{checkPackage}
\title{Check package}
\usage{
checkPackage(
  pkgName,
  repoDir,
  subdir = NULL,
  fail_on = "warning",
  bioc_check = FALSE
)
}
\arguments{
\item{pkgName}{String of package name.}

\item{repoDir}{String of path to repository directory.}

\item{subdir}{String of subpath to package directory.}

\item{fail_on}{String specifying the level at which check fail.}

\item{bioc_check}{Logical whether bioc check should be performed
Supported values: `note`, `warning`(default) and `error`}
}
\description{
Check package
}
