% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{checkPackage}
\alias{checkPackage}
\title{Check package}
\usage{
checkPackage(pkgName, repoDir, subdir)
}
\arguments{
\item{pkgName}{String of package name.}

\item{repoDir}{String of path to repository directory.}

\item{pkgDir}{String of path to package directory.}
}
\description{
Check package
}
