% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRstyle-package.R
\docType{package}
\name{gDRstyle-package}
\alias{gDRstyle}
\alias{gDRstyle-package}
\title{gDRstyle: A package with style requirements for the gDR suite}
\value{
package help page
}
\description{
Package fills a helper package role for whole gDR suite. It helps to support good development practices by keeping style requirements and style tests for other packages. It also contains build helpers to make all package requirements met.
}
\note{
To learn more about functions start with \code{help(package = "gDRstyle")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gdrplatform/gDRstyle}
  \item \url{https://gdrplatform.github.io/gDRstyle/}
  \item Report bugs at \url{https://github.com/gdrplatform/gDRstyle/issues}
}

}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} (\href{https://orcid.org/0000-0002-7059-6378}{ORCID})

Authors:
\itemize{
  \item Allison Vuong
  \item Dariusz Scigocki
  \item Marcin Kamianowski
  \item Janina Smola
  \item Bartosz Czech (\href{https://orcid.org/0000-0002-9908-3007}{ORCID})
}

Other contributors:
\itemize{
  \item Aleksander Chlebowski [contributor]
}

}
\keyword{internal}
