% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gDRstyle-package.R
\docType{package}
\name{gDRstyle-package}
\alias{gDRstyle}
\alias{gDRstyle-package}
\title{gDRstyle: A package with style requirements for the gDR suite}
\value{
package help page
}
\description{
Package fills a helper package role for whole gDR suite. It helps to support good development practices by keeping style requirements and style tests for other packages. It also contains build helpers to make all package requirements met.
}
\note{
To learn more about functions start with \code{help(package = "gDRstyle")}
}
\author{
\strong{Maintainer}: Arkadiusz Gladki \email{gladki.arkadiusz@gmail.com} [contributor]

Authors:
\itemize{
  \item Allison Vuong
  \item Dariusz Scigocki
  \item Marcin Kamianowski
}

Other contributors:
\itemize{
  \item Aleksander Chlebowski [contributor]
  \item Bartosz Czech [contributor]
}

}
\keyword{internal}
