% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{installAllDeps}
\alias{installAllDeps}
\title{Install all package dependencies from yaml file for
building image purposes}
\usage{
installAllDeps(
  additionalRepos = NULL,
  base_dir = "/mnt/vol",
  use_ssh = FALSE,
  test_mode = FALSE
)
}
\arguments{
\item{additionalRepos}{List of additional Repos}

\item{base_dir}{String of base working directory.}

\item{use_ssh}{logical, if use ssh keys}

\item{test_mode}{logical, whether to run the function in the test mode
(if TRUE the dependencies are not installed but only listed)}
}
\value{
\code{NULL}
}
\description{
Install all package dependencies from yaml file for
building image purposes
}
\examples{
installAllDeps(
  base_dir = system.file(package = "gDRstyle", "testdata"),
  test_mode = TRUE
)

}
