% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{installAllDeps}
\alias{installAllDeps}
\title{Install all package dependencies from yaml file for builiding image purposes}
\usage{
installAllDeps(additionalRepos = NULL, base_dir = "/mnt/vol", use_ssh = FALSE)
}
\arguments{
\item{base_dir}{String of base working directory.}

\item{use_ssh}{logical, if use ssh keys}
}
\description{
Install all package dependencies from yaml file for builiding image purposes
}
