% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tools.R
\name{installLocalPackage}
\alias{installLocalPackage}
\title{Install locally cloned repo for builiding image purposes}
\usage{
installLocalPackage(repo_path, additionalRepos = NULL, base_dir = "/mnt/vol")
}
\arguments{
\item{repo_path}{String of repository directory.}

\item{additionalRepos}{List of additional Repos}

\item{base_dir}{String of base working directory.}
}
\value{
\code{NULL}
}
\description{
Install locally cloned repo for builiding image purposes
}
\examples{
installLocalPackage(
  repo_path = "gDRstyle/tst_pkgs/dummy_pkg"
)
remove.packages("fakePkg")

}
