% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linter_custom.R
\name{ticket_ref_linter}
\alias{ticket_ref_linter}
\title{ticket_ref_linter}
\usage{
ticket_ref_linter()
}
\value{
linter class function
}
\description{
Flag Jira ticket references such as \code{GDR-<id>} left in the source
(in code, comments, or strings). The ticket ID belongs in the branch name
only, never in the committed source. A reference inside a \code{TODO} or
\code{FIXME} comment is allowed, since it legitimately links deferred work
to the ticket that tracks it.
}
\examples{
linters_config <- lintr::linters_with_defaults(
  ticket_ref_linter = ticket_ref_linter()
)

}
\author{
Bartosz Czech \href{mailto:bartosz.czech@contractors.roche.com}{bartosz.czech@contractors.roche.com}
}
\keyword{linter}
